requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiogram==3.13.0
python-dotenv==1.0.1
pandas==2.2.0
xlrd==2.0.1
openpyxl==3.1.5
reportlab==4.2.0
aiosqlite==0.20.0
Pillow==10.4.0
python-multipart==0.0.9
